Complete reworked FLARM device driver with IGC download.
