+++ Common/Data/Language/ENG_MSG.TXT
@@ -2354,3 +2354,15 @@ _@M2400_ "Downloading"
 _@M2401_ "No Device found"
 _@M2402_ "Abort Download"
 _@M2403_ "Reset FLARM"
+_@M2404_ "Download"
+_@M2405_ "Own FLARM Id:"
+
+		 
+_@M2406_ "IGC File download complete!"
+_@M2407_ "Error: receive timeout!"
+_@M2408_ "Error: CRC checksum fail!"
+_@M2409_ "Error: invalid filename!"
+_@M2410_ "Error: can't open file!"
+_@M2411_ "Error: communication timeout!"
+_@M2412_ "Error: unknown"
+_@M2413_ "FLARM need reboot for normal operation\n reboot now?"
