+++ Common/Header/dlgFlarmIGCDownload.h
@@ -1,0 +1,18 @@
+/*
+   LK8000 Tactical Flight Computer -  WWW.LK8000.IT
+   Released under GNU/GPL License v.2
+   See CREDITS.TXT file for authors and copyrights
+
+   $Id$
+*/
+
+#ifndef _DLGFLARMIGCDOWNLOAD_H_
+#define _DLGFLARMIGCDOWNLOAD_H_
+
+ListElement* dlgIGCSelectListShowModal(  DeviceDescriptor_t *d) ;
+void StartIGCReadThread();
+void StopIGCReadThread(void) ;
+
+#endif
+
+
